+++ pkg/genericapiserver/BUILD
@@ -23,7 +23,6 @@ go_library(
         "resource_encoding_config.go",
         "serve.go",
         "storage_factory.go",
-        "tunneler.go",
     ],
     tags = ["automanaged"],
     deps = [
@@ -42,11 +41,8 @@ go_library(
         "//pkg/genericapiserver/mux:go_default_library",
         "//pkg/genericapiserver/options:go_default_library",
         "//pkg/genericapiserver/routes:go_default_library",
-        "//pkg/ssh:go_default_library",
         "//pkg/storage/storagebackend:go_default_library",
-        "//pkg/util:go_default_library",
         "//pkg/util/cert:go_default_library",
-        "//pkg/util/clock:go_default_library",
         "//pkg/util/config:go_default_library",
         "//pkg/version:go_default_library",
         "//vendor:github.com/coreos/go-systemd/daemon",
@@ -56,7 +52,6 @@ go_library(
         "//vendor:github.com/golang/glog",
         "//vendor:github.com/pborman/uuid",
         "//vendor:github.com/pkg/errors",
-        "//vendor:github.com/prometheus/client_golang/prometheus",
         "//vendor:gopkg.in/natefinch/lumberjack.v2",
         "//vendor:k8s.io/apimachinery/pkg/apimachinery",
         "//vendor:k8s.io/apimachinery/pkg/apimachinery/registered",
@@ -69,7 +64,6 @@ go_library(
         "//vendor:k8s.io/apimachinery/pkg/util/runtime",
         "//vendor:k8s.io/apimachinery/pkg/util/sets",
         "//vendor:k8s.io/apimachinery/pkg/util/validation",
-        "//vendor:k8s.io/apimachinery/pkg/util/wait",
         "//vendor:k8s.io/apiserver/pkg/authentication/authenticator",
         "//vendor:k8s.io/apiserver/pkg/authentication/request/union",
         "//vendor:k8s.io/apiserver/pkg/authentication/user",
@@ -88,7 +82,6 @@ go_test(
         "serve_test.go",
         "server_run_options_test.go",
         "storage_factory_test.go",
-        "tunneler_test.go",
     ],
     library = ":go_default_library",
     tags = ["automanaged"],
@@ -107,7 +100,6 @@ go_test(
         "//pkg/storage/etcd/testing:go_default_library",
         "//pkg/storage/storagebackend:go_default_library",
         "//pkg/util/cert:go_default_library",
-        "//pkg/util/clock:go_default_library",
         "//pkg/util/config:go_default_library",
         "//pkg/version:go_default_library",
         "//vendor:github.com/go-openapi/spec",
