+++ pkg/master/tunneler/ssh_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package genericapiserver
+package tunneler
 
 import (
 	"fmt"
